Core command host for Arduino devices
Implementation of a serial oriented command interface for arduino devices.

It provides a basic command host which allows introspection of functions
attributes, the ability to set and get attributes, and provision of help.

To use this in your project, you include it, and subclass the command host
and implement the bits that matter to you.