Massive re-org for packaging arduino parts
Lots of structural changes. Library code into "libraries".
Examples into a subdirectory of that. Examples renamed to match
the subdirectories.

This also includes initial basic debian/ubuntu packaging that places
the code in the same place on the system as the other examples/libraries
meaning the libraries/examples are now usable directly from the normal
arduino IDE.