Start of new Stdio based code tester
The idea of the stdio based code tester is that it allows you to compile
arduino targetted code using normal tools, test the code on a local machine
against mock versions of facilities to check that the code does what you
think it should. This particular set of changes is to allow a sketch to use
stdin/stdout as serial in/out comms to/from the arduino to allow behavioural
based tests as well.

An overview of the changes:

  * Clone of MockSerial to develop stdio based version
  * Renamed class/etc to match StdioMockSerial
  * Conditionally include the correct MockSerial based on -D variable
  * Added overview docs for each method
      The idea is who calls what and why, to enable the rewrite for using
      stdio as the source/destination for data. May be able to avoid wholesale
      changes to the mock structure and push out changes to the caller/running
      host. (This would also enable a bunch of automated tests...)
  * Fork the host_test.cpp for developing serial version
  * Added spec for mocking serial data from the client
  * Add stub implementations of serial reading funcs
  * Mass reformat - remove indentation primarily.
  * Harness for testing the serial tester...
  * StdioMockSerial::print implementation started, tested
  * stdio serial based tests passing.