Creation of the TestHost and device oriented fixes
The TestHost was the first host to be run (successfully) on a device
that was actually the TestHost - rather than something small and minimalistic.
One of the things it showed up is that the original "return a string" style
API had memory issues on real devices - even the Atmega 328 failed to have
sufficient memory to run the API completely for this - even though the Atmega8
handled things like the BotHost reasonably well. This is due to the sheer amount
of memory used in the highly constrained environment. As a result, the
code was changed to the form found in this commit and to fit inside a real world
device.

Initial versions of the TestHost

BUGFIX: Changes to allow working on actual devices.

Return 404 for unimplemented things which are called

Redo to fit with changed core API

Change example to fit changed device internal API