Retired OvenSpike and Impractical CommandHost
The Ovenspike was the original demo idea, but a) isn't very exciting
and b) doesn't relate to a physical device. It's therefore been retired.

The arduino CommandHost likewise originally used arduino String objects
but these eat up too much memory on a Atmega8 and as a result char* needs
to be used there. Originally I considered having 2 versions (a non-tiny and
tiny version), but after extensive work, this looks like it would be too
much to maintain in the long term, too confusing for new users and also
lead to confusion about memory usage issues. As a result that's retired too.

For now, both have been shifted into an "attic" allowing browsing of historical
stuff for convenience, but otherwise no longer used and deprecated.